PPM_1: unitN now has lateral inhibition, but problematic
unitN.m now has N-way lateral inhibition and everything is equivalent to
the things in unit2.m. However, the result for N-unit result is pretty bad
when N is larger than 3 or 5. I have spotted an error in the differential
equation solver. Taking a snapshot before I proceed to correcting that
error.